+++ src/charmer/__init__.py
@@ -1,3 +1,3 @@
 """Charmer Pangolin provisioner."""
 
-__version__ = "0.9.0"
+__version__ = "0.10.0"